CircleAI 3.3.0 — HER/JARVIS parity — always-on, world-aware, embodied, learns-you, reasoner [skip ci]
28 deliverables across 5 phases against existing HER/Jarvis contracts:
- A (always-on): KWS wake word, Android foreground + iOS background services, conversation handoff
- B (world-aware): Google/MS-Graph/CalDAV calendar, IMAP/Gmail/Graph email, RSS/News/Bluesky/Mastodon, Open-Meteo + OSRM, proactive briefing
- C (embodied): HomeAssistant REST bridge, Camera2/AVFoundation/MediaCapture, YOLO+ArcFace ONNX, Health Connect + HealthKit, GPS
- D (learns-you): mnn-bridge 1.4.0 LoRA training surface (training-disabled fallback), nightly trainer, adapter sync
- E (reasoner): reasoning-loop inner monologue, SQLite KG + LLM extractor, HippoRAG PPR, Bayesian world model, n-gram sequence predictor, ECAPA-TDNN speaker ID, wav2vec2 emotion, SelfBench harness + regression gate

Version-aligned 49 runtime-tier packages 3.2.0 → 3.3.0; 7 new (Integration*, SelfBench) join at 3.3.0; touched 1.x bumped one minor (Companion 1.2→1.3, Voice 1.2→1.3, Memory 1.3→1.4). Native libmnnbridge.dylib + libMNN.dylib staged for osx-arm64.

+++ tests/CircleAI.Tests/TurboQuantParityTests.cs
@@ -110,7 +110,7 @@ public sealed class TurboQuantParityTests
         // Box-Muller over a SplitMix64 — same recipe as the codec's own RNG.
         // Used to fabricate a deterministic but realistic-looking input.
         var v = new float[dim];
-        ulong state = (ulong)seed | (ulong)seed << 32;
+        ulong state = (uint)seed | ((ulong)(uint)seed << 32);
         bool haveSpare = false;
         double spare = 0;
         for (int i = 0; i < dim; i++)
